+++ crates/ironrdp-graphics/src/color_conversion.rs
@@ -18,10 +18,6 @@ fn clip(v: i32) -> u8 {
     min(max(v, 0), 255) as u8
 }
 
-fn clip_i16(v: i16) -> u8 {
-    min(max(v, 0), 255) as u8
-}
-
 #[derive(Debug)]
 pub struct YCbCrBuffer<'a> {
     pub y: &'a [i16],
@@ -57,13 +53,6 @@ pub struct YCbCr {
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct YCoCg {
-    pub y: u8,
-    pub co: i8,
-    pub cg: i8,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct Rgb {
     pub r: u8,
     pub g: u8,
@@ -86,22 +75,6 @@ impl From<YCbCr> for Rgb {
         let r = clip((yy.overflowing_add(cr_r).0) >> 21);
         let g = clip((yy.overflowing_sub(cb_g).0.overflowing_sub(cr_g).0) >> 21);
         let b = clip((yy.overflowing_add(cb_b).0.overflowing_add(cr_b).0) >> 21);
-
-        Self { r, g, b }
-    }
-}
-
-impl From<YCoCg> for Rgb {
-    fn from(YCoCg { y, co, cg }: YCoCg) -> Self {
-        let y = i16::from(y);
-        let co = i16::from(co);
-        let cg = i16::from(cg);
-
-        let t = y - cg;
-
-        let r = clip_i16(t + co);
-        let g = clip_i16(y + cg);
-        let b = clip_i16(t - co);
 
         Self { r, g, b }
     }
